requirements.txt
requests>=2.31.0
PyYAML>=6.0
lxml>=4.9.3
pytest>=8.0.0
jsonschema>=4.19.0